Replaced BSD types by stdint types in all app samples.
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@2955 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/app/eeprom/eeprom_test.c
@@ -152,7 +152,7 @@ int main(void)
 
 	uint8_t txBuffer[128];
 	uint8_t rxBuffer[128];
-    u_long baud = 115200;
+    uint32_t baud = 115200;
 
 	/*
      * Register the UART device, open it, assign stdout to it and set
