+++ nut/app/led_key/pbtest.c
@@ -133,7 +133,7 @@ THREAD( Key2Thread, arg)
  */
 int main(void)
 {
-    u_long baud = 115200;
+    uint32_t baud = 115200;
     /*
      * Register the UART device, open it, assign stdout to it and set
      * the baudrate.
